requirements.txt
attrs==20.3.0
boto3==1.16.23
botocore==1.19.23
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
ecdsa==0.14.1
envs==1.3
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
phonenumbers==8.12.21
pyasn1==0.4.8
pycryptodome==3.3.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jose==3.2.0
python-jose-cryptodome==1.3.2
requests==2.25.0
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
urllib3==1.26.4
warrant==0.6.1
Werkzeug==1.0.1
Django~=3.2
pynamodb~=5.0.3
jose~=1.0.0
warrant~=0.6.1
uuid==1.30
